requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
python-multipart
pdf2image
pillow
